+++ src/components/_FunctionButtons.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import InputButton from "./InputButton";
 import {MdFunctions, MdInsertChart} from "react-icons/md";
-import {mdiFunctionVariant} from "@mdi/js";
+import { TbMathFunction } from "react-icons/tb";
 import { type WithTranslation, withTranslation } from "react-i18next";
 
 type FunctionInputButtonsInternalProps = {
@@ -24,9 +24,7 @@ class FunctionInputButtonsInternal extends React.Component<FunctionInputButtonsI
           onClick={this.props.onExpressionClick}
           title={t("Convert to expression")}
         >
-          <svg style={{width:"14px", height:"14px", verticalAlign: "middle"}} viewBox="0 0 24 24">
-            <path fill="currentColor" d={mdiFunctionVariant} />
-          </svg>
+          <TbMathFunction />
         </InputButton>
       );