+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### ✨ Features and improvements
 - Added translation to "Links" in debug modal
 - Add support for hillshade's color arrays and relief-color elevation expression
+- Remove `@mdi` packages in favor of `react-icons`
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes